Fix WAL batch atomicity and strengthen gap detection test
- Modified `src/storage/wal/concurrent.rs`: Added pre-validation loop to `append_batch` to ensure all entries meet size limits before any LSN allocation or writing occurs. This prevents partial writes and LSN gaps.
- Replaced `tests/havoc_wal_gaps.rs` with `tests/wal_batch_atomicity.rs`: The new test asserts that batch failure is atomic (no LSNs consumed) rather than asserting fragility (gaps).
- Updated `.jules/elenchus.md`: Documented the "Test Inversion" finding.
